Creates a directory of file-path if it doesn't exist
Often, when creating new files, the path may not entirely be
created (until the users saves the buffers). This greats in the way of
getting diagnostics for the file.

A good (but not the most user friendly) is to create the directory
ourselves.

A better approach is to recursively climb up until an existent path is
found and then that path is passed to `esy/project--of-path`